creating the correct folder structure for vehicle details and implementing cat c
