+++ src/modules/tests/journal-data/cat-c/candidate/candidate.cat-c.actions.ts
@@ -3,10 +3,10 @@ import { CatCUniqueTypes } from '@dvsa/mes-test-schema/categories/C';
 
 export const POPULATE_CANDIDATE_DETAILS_CAT_C = '[CatC] [JournalEffects] Populate Candidate Details';
 
-export class PopulateCandidateDetailsCatBE implements Action {
+export class PopulateCandidateDetailsCatC implements Action {
   readonly type = POPULATE_CANDIDATE_DETAILS_CAT_C;
   constructor(public payload: CatCUniqueTypes.Candidate) {}
 }
 
 export type Types =
-  | PopulateCandidateDetailsCatBE;
+  | PopulateCandidateDetailsCatC;
