+++ src/modules/tests/journal-data/common/candidate/candidate.action-creator.ts
@@ -1,6 +1,7 @@
 import { TestCategory } from '@dvsa/mes-test-schema/category-definitions/common/test-category';
 import { PopulateCandidateDetails } from '../../cat-b/candidate/candidate.actions';
 import { PopulateCandidateDetailsCatBE } from '../../cat-be/candidate/candidate.cat-be.actions';
+import { PopulateCandidateDetailsCatC } from '../../cat-c/candidate/candidate.cat-c.actions';
 import { Booking } from '@dvsa/mes-journal-schema';
 import { get } from 'lodash';
 
@@ -16,7 +17,7 @@ export const createPopulateCandidateDetailsAction = (testCategory: string, booki
         businessTelephone: get(booking, 'business.telephone'),
       });
     case TestCategory.C:
-      return new PopulateCandidateDetailsCatBE({
+      return new PopulateCandidateDetailsCatC({
         ...booking.candidate,
         businessAddress: get(booking, 'business.businessAddress'),
         businessName: get(booking, 'business.businessName'),
