+++ src/modules/tests/vehicle-details/cat-b/vehicle-details.cat-b.selector.ts
@@ -1,0 +1,4 @@
+import { CatBUniqueTypes } from '@dvsa/mes-test-schema/categories/B/index';
+
+export const getSchoolCar = (vehicleDetails: CatBUniqueTypes.VehicleDetails) => vehicleDetails.schoolCar;
+export const getDualControls = (vehicleDetails: CatBUniqueTypes.VehicleDetails) => vehicleDetails.dualControls;
