+++ src/modules/tests/vehicle-details/cat-be/__tests__/vehicle-details.cat-be.selector.spec.ts
