+++ src/modules/tests/vehicle-details/cat-be/vehicle-details.cat-be.selector.ts
@@ -1,5 +1,6 @@
-// @TODO - import Cat C Types
 import { CatBEUniqueTypes } from '@dvsa/mes-test-schema/categories/BE';
 
 export const getVehicleLength = (vehicleDetails: CatBEUniqueTypes.VehicleDetails) => vehicleDetails.vehicleLength;
 export const getVehicleWidth = (vehicleDetails: CatBEUniqueTypes.VehicleDetails) => vehicleDetails.vehicleWidth;
+export const getSchoolCar = (vehicleDetails: CatBEUniqueTypes.VehicleDetails) => vehicleDetails.schoolCar;
+export const getDualControls = (vehicleDetails: CatBEUniqueTypes.VehicleDetails) => vehicleDetails.dualControls;
