+++ src/modules/tests/vehicle-details/cat-c/vehicle-details.cat-c.selector.ts
@@ -1,0 +1,4 @@
+import { CatCUniqueTypes } from '@dvsa/mes-test-schema/categories/C';
+
+export const getVehicleLength = (vehicleDetails: CatCUniqueTypes.VehicleDetails) => vehicleDetails.vehicleLength;
+export const getVehicleWidth = (vehicleDetails: CatCUniqueTypes.VehicleDetails) => vehicleDetails.vehicleWidth;
