+++ src/modules/tests/vehicle-details/common/vehicle-details.actions.ts
