--- src/modules/tests/vehicle-details/vehicle-details.cat-be.selector.ts
@@ -1,4 +1,0 @@
-import { CatBEUniqueTypes } from '@dvsa/mes-test-schema/categories/BE';
-
-export const getVehicleLength = (vehicleDetails: CatBEUniqueTypes.VehicleDetails) => vehicleDetails.vehicleLength;
-export const getVehicleWidth = (vehicleDetails: CatBEUniqueTypes.VehicleDetails) => vehicleDetails.vehicleWidth;
