+++ src/pages/non-pass-finalisation/cat-c/non-pass-finalisation.cat-c.page.ts
@@ -55,10 +55,10 @@ import { SetTestStatusWriteUp } from '../../../modules/tests/test-status/test-st
 import { SetActivityCode } from '../../../modules/tests/activity-code/activity-code.actions';
 import { BasePageComponent } from '../../../shared/classes/base-page';
 import { getGearboxCategory, isAutomatic, isManual } from
-'../../../modules/tests/vehicle-details/vehicle-details.selector';
-import { getVehicleDetails } from '../../../modules/tests/vehicle-details/vehicle-details.reducer';
+'../../../modules/tests/vehicle-details/common/vehicle-details.selector';
+import { getVehicleDetails } from '../../../modules/tests/vehicle-details/cat-b/vehicle-details.reducer';
 import { GearboxCategory } from '@dvsa/mes-test-schema/categories/common';
-import { GearboxCategoryChanged } from '../../../modules/tests/vehicle-details/vehicle-details.actions';
+import { GearboxCategoryChanged } from '../../../modules/tests/vehicle-details/common/vehicle-details.actions';
 
 interface NonPassFinalisationPageState {
   candidateName$: Observable<string>;
