+++ src/pages/waiting-room-to-car/cat-be/waiting-room-to-car.cat-be.page.ts
@@ -12,7 +12,7 @@ import {
   DualControlsToggled,
   GearboxCategoryChanged,
   VehicleRegistrationChanged,
-} from '../../../modules/tests/vehicle-details/vehicle-details.actions';
+} from '../../../modules/tests/vehicle-details/common/vehicle-details.actions';
 import { map } from 'rxjs/operators';
 import {
   InstructorAccompanimentToggled,
@@ -20,16 +20,18 @@ import {
   SupervisorAccompanimentToggled,
   InterpreterAccompanimentToggled,
 } from '../../../modules/tests/accompaniment/accompaniment.actions';
-import { getVehicleDetails } from '../../../modules/tests/vehicle-details/vehicle-details.reducer';
+import { getVehicleDetails } from '../../../modules/tests/vehicle-details/cat-b/vehicle-details.reducer';
 import { getAccompaniment } from '../../../modules/tests/accompaniment/accompaniment.reducer';
 import {
   getRegistrationNumber,
   getGearboxCategory,
-  getSchoolCar,
-  getDualControls,
   isAutomatic,
   isManual,
-} from '../../../modules/tests/vehicle-details/vehicle-details.selector';
+} from '../../../modules/tests/vehicle-details/common/vehicle-details.selector';
+import {
+  getSchoolCar,
+  getDualControls,
+} from '../../../modules/tests/vehicle-details/cat-be/vehicle-details.cat-be.selector';
 import {
   getInstructorAccompaniment,
   getSupervisorAccompaniment,
